TAG105 2019/08/22   Core     ・YM2151/K051649/HuC6280のベンド処理修正。  ・実データのないXGMを作成しようとすると失敗するバグを修正。
  mml2vgmIDE
    ・F5を連打するとクラッシュするバグを修正。
 ・設定から実チップを使用しないことができるようにした。
 ・コンパイルエラーが１行目で発生した時、そこにキャレットを移動しようとするとクラッシュするバグを修正。

  発見したけど残ってる不具合(増えたorz)
    form1 タスクマネージャに出てこないようにしたい
 未使用音源のエンベープを定義すると発生する警告
 tonedoubler 動作しない
 conductor パラメータ表示不具合
 dcsgSecondary パラメータ表示不具合
 HuC6280 PCMMap

+++ mml2vgm/Core/chips/AY8910.cs
@@ -351,7 +351,7 @@ namespace Core
         {
             if (mml.args[0] is string) return;
 
-            int adr = (int)mml.args[0];
+            int adr = (byte)mml.args[0];
             byte dat = (byte)mml.args[1];
 
             parent.OutData(mml, port[0], (byte)adr, (byte)dat);
